requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
pyjwt==2.8.0

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.6.0
redis==7.1.0

# API & LLM
anthropic==0.18.1
openai==1.12.0

# Translation
google-cloud-translate==3.15.0
langdetect==1.0.9

# Web Crawling
scrapy==2.11.0
playwright==1.41.0
selenium==4.16.0
newspaper3k==0.2.8
beautifulsoup4==4.14.3
lxml==5.3.0
lxml_html_clean==0.4.3
aiohttp==3.9.1

# Utilities
requests==2.32.5
python-dateutil==2.8.2
pytz==2024.1

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1

# Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-celery==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-exporter-otlp==1.27.0
prometheus-client==0.21.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
